requirements.txt
absl-py==2.4.0
birder==0.4.2
braceexpand==0.1.7
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.3
fonttools==4.61.1
fsspec==2026.2.0
grpcio==1.76.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lightning-utilities==0.15.2
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
onnx==1.20.1
onnx-ir==0.1.15
onnxscript==0.6.0
packaging==26.0
piexif==1.1.3
pillow==12.1.0
polars==1.38.0
polars-runtime-32==1.38.0
protobuf==6.33.5
pyarrow==23.0.0
pycocotools==2.0.11
Pygments==2.19.2
pyparsing==3.3.2
PyQt6==6.10.2
PyQt6-Qt6==6.10.2
PyQt6-WebEngine==6.10.0
PyQt6-WebEngine-Qt6==6.10.2
PyQt6_sip==13.11.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
rich==14.3.2
scikit-learn==1.8.0
scipy==1.17.0
setuptools==80.10.2
six==1.17.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
torch==2.10.0
torchinfo==1.8.0
torchmetrics==1.8.2
torchvision==0.25.0
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
webdataset==1.0.2
Werkzeug==3.1.5
